Restructure the files     * Move the Dao.java to the dao package     * Move the DatabaseConnector.java to the utils package     * Update the import statements in the Main.java and DatabaseConnectorTest.java
