+++ src/main/java/utils/DatabaseConnector.java
@@ -1,4 +1,4 @@
-package dao;
+package utils;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
