+++ src/test/java/utils/DatabaseConnectorTest.java
@@ -1,5 +1,6 @@
-package dao;
+package utils;
 
+import utils.DatabaseConnector;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
